workbench copy 2/requirements.txt
pydantic-ai
python-dotenv
httpx
logfire
spotipy
PyGithub